+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * transform from ide docker image to dojo docker image #17139
 * no need for custom entrypoint - set java variables in etc/dojo.d/variables/
+* pretty bash prompt
 
 ### 0.5.0 (2018-Sep-10)